Split out test discovery into its own file.
