+++ smoke.cabal
@@ -18,7 +18,8 @@ source-repository head
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Test.Smoke Test.Smoke.FileTypes
+  exposed-modules:     Test.Smoke
+  other-modules:       Test.Smoke.Discovery Test.Smoke.FileTypes Test.Smoke.Types
   default-language:    Haskell2010
   build-depends:       base >= 4.9.1 && < 5
                      , directory
